move subscription configuration into subscriptions area
