+++ src/Beckett/ServiceCollectionExtensions.cs
@@ -1,7 +1,4 @@
-using Beckett.Configuration;
-using Beckett.Dashboard;
 using Beckett.Database;
-using Beckett.MessageStorage;
 using Beckett.OpenTelemetry;
 using Beckett.Scheduling;
 using Beckett.Storage;
@@ -34,8 +31,6 @@ public static class ServiceCollectionExtensions
         services.AddSingleton(options);
 
         services.AddSingleton<IMessageStore, MessageStore>();
-
-        services.AddDashboardSupport();
 
         services.AddMessageStorageSupport(options);
