+++ src/Beckett/Subscriptions/Configuration/SubscriptionConfigurationBuilder.cs
@@ -1,7 +1,6 @@
-using Beckett.Subscriptions;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Beckett.Configuration;
+namespace Beckett.Subscriptions.Configuration;
 
 public class SubscriptionConfigurationBuilder(
     Subscription subscription,
